mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.52.1",
    "mcp>=1.9.1",
    "python-dotenv>=1.1.0",
]
